// frontend/src/client/models/GeneratePageResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * Response from generation trigger
 */
export type GeneratePageResponse = {
    generation_version_id: string;
    page_id: string;
    website_id: string;
    subdomain: string;
    status: string;
    message?: string;
};